fixing inspiration table not found on production DB

+++ migrations/20230831060337_createInspirationTable.js
@@ -3,15 +3,19 @@
  * @returns { Promise<void> }
  */
 exports.up = function(knex) {
-    return knex.schema.createTable('inspiration', (table) => {
-        table.increments('id').primary();
-        table.integer('user_id')
-        .unsigned()
-        .references('users.id')
-        .onUpdate('CASCADE')
-        .onDelete('CASCADE');
-        table.string('item_name').notNullable();
-    })
+    return knex.schema.hasTable('inspiration').then(exists => {
+        if (!exists) {
+            return knex.schema.createTable('inspiration', (table) => {
+                table.increments('id').primary();
+                table.integer('user_id')
+                .unsigned()
+                .references('users.id')
+                .onUpdate('CASCADE')
+                .onDelete('CASCADE');
+                table.string('item_name').notNullable();
+            });
+        }
+    });
 };
 
 /**
@@ -19,5 +23,9 @@ exports.up = function(knex) {
  * @returns { Promise<void> }
  */
 exports.down = function(knex) {
-    return knex.schema.dropTable('inspiration');
+    return knex.schema.hasTable('inspiration').then(exists => {
+        if (exists) {
+            return knex.schema.dropTable('inspiration');
+        }
+    });
 };
